+++ firmware/controllers/can/rusefi_wideband.cpp
@@ -4,6 +4,9 @@
 
 #include "ch.h"
 #include "can_msg_tx.h"
+#include "rusefi_wideband.h"
+#include "sensor.h"
+#include "engine.h"
 
 // This file contains an array called build_wideband_noboot_bin
 // This array contains the firmware image for the wideband contoller
@@ -117,6 +120,17 @@ void setWidebandOffset(uint8_t index) {
 	}
 
 	waitingBootloaderThread = nullptr;
+}
+
+void sendWidebandInfo() {
+	CanTxMessage m(0xEF5'0000, 2, true);
+
+	float vbatt = Sensor::get(SensorType::BatteryVoltage).value_or(0) * 10;
+
+	m[0] = vbatt;
+
+	// Offset 1 bit 0 = heater enable
+	m[1] = enginePins.o2heater.getLogicValue() ? 0x01 : 0x00;
 }
 
 #endif // EFI_WIDEBAND_FIRMWARE_UPDATE && HAL_USE_CAN