+++ firmware/controllers/lua/lua.cpp
@@ -16,8 +16,6 @@
 #include "engine.h"
 #include "tunerstudio_outputs.h"
 
-EXTERN_ENGINE;
-
 #define LUA_HEAP_SIZE 20000
 
 static memory_heap_t heap;
@@ -368,6 +366,18 @@ int testLuaReturnsInteger(const char* script) {
 	}
 
 	return lua_tointeger(ls, -1);
+}
+
+void testLuaExecString(const char* script) {
+	auto ls = setupLuaState();
+
+	if (!ls) {
+		throw new std::logic_error("Call to setupLuaState failed, returned null");
+	}
+
+	if (!loadScript(ls, script)) {
+		throw new std::logic_error("Call to loadScript failed");
+	}
 }
 
 #endif // EFI_UNIT_TEST