+++ firmware/controllers/system/timer/pwm_generator_logic.cpp
@@ -13,8 +13,6 @@
 #include "pwm_generator_logic.h"
 #include "perf_trace.h"
 
-EXTERN_ENGINE;
-
 #if EFI_PROD_CODE
 #include "mpu_util.h"
 #include "engine.h"
@@ -269,7 +267,7 @@ static void timerCallback(PwmConfig *state) {
 		return;
 	}
 
-	state->executor->scheduleByTimestampNt(&state->scheduling, switchTimeNt, { timerCallback, state });
+	state->executor->scheduleByTimestampNt("pwm", &state->scheduling, switchTimeNt, { timerCallback, state });
 	state->dbgNestingLevel--;
 }