+++ firmware/init/sensor/init_lambda.cpp
@@ -6,8 +6,6 @@
 #include "function_pointer_sensor.h"
 #include "ego.h"
 
-EXTERN_ENGINE;
-
 struct GetAfrWrapper {
 	DECLARE_ENGINE_PTR;
 
@@ -34,6 +32,11 @@ void initLambda(DECLARE_ENGINE_PARAMETER_SIGNATURE) {
 
 #if EFI_CAN_SUPPORT
 	if (CONFIG(enableAemXSeries)) {
+		if (!CONFIG(canWriteEnabled) || !CONFIG(canReadEnabled)) {
+			firmwareError(OBD_PCM_Processor_Fault, "CAN read and write are required to use CAN wideband.");
+			return;
+		}
+
 		registerCanSensor(aem1);
 		registerCanSensor(aem2);